Struct calyx_opt::pass_manager::PassManager

pub struct PassManager { /* private fields */ }

Structure that tracks all registered passes for the compiler.

Implementations

impl PassManager

pub fn default_passes() -> PassResult<Self>

impl PassManager

pub fn register_pass<Pass>(&mut self) -> PassResult<()>
where Pass: Visitor + ConstructVisitor + Named,

Register a new Calyx pass and return an error if another pass with the same name has already been registered.

Example

let pm = PassManager::default();
pm.register_pass::<WellFormed>()?;

pub fn register_diagnostic<Pass>(&mut self) -> PassResult<()>
where Pass: Visitor + ConstructVisitor + Named + DiagnosticPass,

Registers a diagnostic pass as a normal pass. If there is an error, this will report the first error gathered by the pass.

pub fn add_alias(&mut self, name: String, passes: Vec<String>) -> PassResult<()>

Adds a new alias for groups of passes. An alias is a list of strings that represent valid pass names OR an alias. The passes and aliases are executed in the order of specification.

pub fn specific_help(&self, pass: &str) -> Option<String>

Return the help string for a specific pass.

pub fn complete_help(&self) -> String

Return a string representation to show all available passes and aliases. Appropriate for help text.

pub fn execute_plan( &self, ctx: &mut Context, incl: &[String], excl: &[String], insn: &[String], dump_ir: bool, ) -> PassResult<()>

Executes a given “plan” constructed using the incl and excl lists. ord is a relative ordering that should be enforced.