Struct calyx_opt::passes::DiscoverExternal

pub struct DiscoverExternal { /* private fields */ }

A pass to detect cells that have been inlined into the top-level component and turn them into real cells marked with ir::BoolAttr::External.