Trait calyx_opt::traversal::Named

pub trait Named {
    // Required methods
    fn name() -> &'static str;
    fn description() -> &'static str;

    // Provided method
    fn opts() -> Vec<PassOpt> { ... }
}

Trait that describes named things. Calling do_pass and do_pass_default. require this to be implemented.

This has to be a separate trait from Visitor because these methods don’t recieve self which means that it is impossible to create dynamic trait objects.

Required Methods

fn name() -> &'static str

The name of a pass. Is used for identifying passes.

fn description() -> &'static str

A short description of the pass.

Provided Methods

fn opts() -> Vec<PassOpt>

Set of options that can be passed to the pass. The options contains a tuple of the option name and a description.

Object Safety

This trait is not object safe.