Trait calyx_opt::traversal::Named

pub trait Named {
    fn name() -> &'static str;
    fn description() -> &'static str;

    fn opts() -> &'static [(&'static str, &'static str)] { ... }
}

Trait that describes named things. Calling do_pass and do_pass_default. require this to be implemented.

This has to be a separate trait from Visitor because these methods don’t recieve self which means that it is impossible to create dynamic trait objects.

Required Methods

fn name() -> &'static str

The name of a pass. Is used for identifying passes.

fn description() -> &'static str

A short description of the pass.

Provided Methods

fn opts() -> &'static [(&'static str, &'static str)]

Set of options that can be passed to the pass. The options contains a tuple of the option name and a description.