Trait calyx_opt::traversal::Visitor

pub trait Visitor {

    fn precondition(_ctx: &Context) -> Option<String>
    where
        Self: Sized,
    { ... }
    fn start_context(&mut self, _ctx: &mut Context) -> VisResult { ... }
    fn finish_context(&mut self, _ctx: &mut Context) -> VisResult { ... }
    fn iteration_order() -> Order
    where
        Self: Sized,
    { ... }
    fn traverse_component(
        &mut self,
        comp: &mut Component,
        signatures: &LibrarySignatures,
        components: &[Component]
    ) -> CalyxResult<()>
    where
        Self: Sized,
    { ... }
    fn do_pass(&mut self, context: &mut Context) -> CalyxResult<()>
    where
        Self: Sized + ConstructVisitor + Named,
    { ... }
    fn do_pass_default(context: &mut Context) -> CalyxResult<Self>
    where
        Self: ConstructVisitor + Sized + Named,
    { ... }
    fn start(
        &mut self,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish(
        &mut self,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_seq(
        &mut self,
        _s: &mut Seq,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_seq(
        &mut self,
        _s: &mut Seq,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_par(
        &mut self,
        _s: &mut Par,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_par(
        &mut self,
        _s: &mut Par,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_if(
        &mut self,
        _s: &mut If,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_if(
        &mut self,
        _s: &mut If,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_while(
        &mut self,
        _s: &mut While,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_while(
        &mut self,
        _s: &mut While,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_repeat(
        &mut self,
        _s: &mut Repeat,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_repeat(
        &mut self,
        _s: &mut Repeat,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_static_control(
        &mut self,
        _s: &mut StaticControl,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_static_control(
        &mut self,
        _s: &mut StaticControl,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn enable(
        &mut self,
        _s: &mut Enable,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn static_enable(
        &mut self,
        _s: &mut StaticEnable,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_static_if(
        &mut self,
        _s: &mut StaticIf,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_static_if(
        &mut self,
        _s: &mut StaticIf,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_static_repeat(
        &mut self,
        _s: &mut StaticRepeat,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_static_repeat(
        &mut self,
        _s: &mut StaticRepeat,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_static_seq(
        &mut self,
        _s: &mut StaticSeq,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_static_seq(
        &mut self,
        _s: &mut StaticSeq,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn start_static_par(
        &mut self,
        _s: &mut StaticPar,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn finish_static_par(
        &mut self,
        _s: &mut StaticPar,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn invoke(
        &mut self,
        _s: &mut Invoke,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn static_invoke(
        &mut self,
        _s: &mut StaticInvoke,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }
    fn empty(
        &mut self,
        _s: &mut Empty,
        _comp: &mut Component,
        _sigs: &LibrarySignatures,
        _comps: &[Component]
    ) -> VisResult { ... }

}

The visiting interface for a ir::Control program. Contains two kinds of functions:

start_: Called when visiting top-down.

finish_: Called when visiting bottow-up.

A pass will usually override one or more function and rely on the default visitors to automatically visit the children.

Provided Methods

fn precondition(_ctx: &Context) -> Option<String>where
Self: Sized,

Precondition for this pass to run on the program. If this function returns None, the pass triggers. Otherwise it aborts and logs the string as the reason.

fn start_context(&mut self, _ctx: &mut Context) -> VisResult

Transform the ir::Context before visiting the components.

fn finish_context(&mut self, _ctx: &mut Context) -> VisResult

Transform the ir::Context after visiting the components.

fn iteration_order() -> Orderwhere
Self: Sized,

Define the iteration order in which components should be visited

fn traverse_component(
&mut self,
comp: &mut Component,
signatures: &LibrarySignatures,
components: &[Component]
) -> CalyxResult<()>where
Self: Sized,

Define the traversal over a component. Calls Visitor::start, visits each control node, and finally calls Visitor::finish.

fn do_pass(&mut self, context: &mut Context) -> CalyxResult<()>where
Self: Sized + ConstructVisitor + Named,

Run the visitor on a given program ir::Context. The function mutably borrows the control program in each component and traverses it.

After visiting a component, it called ConstructVisitor::clear_data to reset the struct.

Panics

Panics if the pass attempts to use the control program mutably.

fn do_pass_default(context: &mut Context) -> CalyxResult<Self>where
Self: ConstructVisitor + Sized + Named,

Build a Default implementation of this pass and call Visitor::do_pass using it.

fn start(
&mut self,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before the traversal begins.

fn finish(
&mut self,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after the traversal ends. This method is always invoked regardless of the Action returned from the children.

fn start_seq(
&mut self,
_s: &mut Seq,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::Seq node.

fn finish_seq(
&mut self,
_s: &mut Seq,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::Seq node.

fn start_par(
&mut self,
_s: &mut Par,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::Par node.

fn finish_par(
&mut self,
_s: &mut Par,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::Par node.

fn start_if(
&mut self,
_s: &mut If,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::If node.

fn finish_if(
&mut self,
_s: &mut If,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::If node.

fn start_while(
&mut self,
_s: &mut While,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::While node.

fn finish_while(
&mut self,
_s: &mut While,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::While node.

fn start_repeat(
&mut self,
_s: &mut Repeat,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::Repeat node.

fn finish_repeat(
&mut self,
_s: &mut Repeat,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::Repeat node.

fn start_static_control(
&mut self,
_s: &mut StaticControl,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the contents of an ir::StaticControl node.

fn finish_static_control(
&mut self,
_s: &mut StaticControl,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the conetnts of an ir::StaticControl node.

fn enable(
&mut self,
_s: &mut Enable,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed at an ir::Enable node.

fn static_enable(
&mut self,
_s: &mut StaticEnable,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed at an ir::StaticEnable node.

fn start_static_if(
&mut self,
_s: &mut StaticIf,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::StaticIf node.

fn finish_static_if(
&mut self,
_s: &mut StaticIf,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::StaticIf node.

fn start_static_repeat(
&mut self,
_s: &mut StaticRepeat,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed before visiting the children of a ir::StaticRepeat node.

fn finish_static_repeat(
&mut self,
_s: &mut StaticRepeat,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed after visiting the children of a ir::StaticRepeat node.

fn start_static_seq(
&mut self,
_s: &mut StaticSeq,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

fn finish_static_seq(
&mut self,
_s: &mut StaticSeq,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

fn start_static_par(
&mut self,
_s: &mut StaticPar,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

fn finish_static_par(
&mut self,
_s: &mut StaticPar,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

fn invoke(
&mut self,
_s: &mut Invoke,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed at an ir::Invoke node.

fn static_invoke(
&mut self,
_s: &mut StaticInvoke,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed at a ir::StaticInvoke node.

fn empty(
&mut self,
_s: &mut Empty,
_comp: &mut Component,
_sigs: &LibrarySignatures,
_comps: &[Component]
) -> VisResult

Executed at an ir::Empty node.